+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 2.0.0-0 / Unreleased
   * Drop dependency on jQuery
   * Drop support for IE
+  * Switch to TypeScript
 
 # 1.3.1-0 / Unreleased